Made datalayer folder and started restructuring model
