+++ app/src/main/res/layout/fragment_log_data.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".model.DB.LogData_frag">
+    tools:context=".datalayer.sheets_db.LogData_frag">
 
 
 </FrameLayout>
